# openapi/patients-gui.yml
openapi: 3.0.3
info:
    title: Patient Registry
    description: 'This is a backend of Patient Registry.'
    termsOfService: TODO
    contact:
        email: jacek.sztajnke@gmail.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    version: 1.0.0
externalDocs:
    description: Find out more about Swagger
    url: http://swagger.io
servers:
    -   url: https://petstore.swagger.io/v1
    -   url: http://petstore.swagger.io/v1
tags:
    -   name: patient
paths:
    /patient:
        put:
            tags:
                -   patient
            summary: Update an existing patient
            operationId: updatePatient
            requestBody:
                description: Patient object that needs to be added to the registry
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Patient'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Patient'
                required: true
            responses:
                400:
                    description: Invalid ID supplied
                    content: {}
                404:
                    description: Patient not found
                    content: {}
                405:
                    description: Validation exception
                    content: {}
#            security:
#            -   patient_registry_auth:
#                    - write:patients
#                    - read:patients
#            x-codegen-request-body-name: body
        post:
            tags:
                - patient
            summary: Add a new patient to the registry
            operationId: addPatient
            requestBody:
                description: Patient object that needs to be added to the registry
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Patient'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Patient'
                required: true
            responses:
                405:
                    description: Invalid input
                    content: {}
#            security:
#            -   patient_registry_auth:
#                    - write:patients
#                    - read:patients
#            x-codegen-request-body-name: body

components:
    schemas:
        Patient:
            type: object
            properties:
                id:
                    type: string
                firstNames:
                    type: string
                lastName:
                    type: string
                pesel:
                    type: string
                birthdate:
                    type: string
                    format: date-time
                status:
                    type: string
                    description: Patients status
                    enum:
                        - new
                        - accepted
                        - deleted
                checksum:
                    type: string

#    securitySchemes:
#        patient_registry_auth:
#            type: oauth2
#            flows:
#                implicit:
#                    authorizationUrl: http://petstore.swagger.io/oauth/dialog
#                    scopes:
#                        write:pets: modify pets in your account
#                        read:pets: read your pets
#        api_key:
#            type: apiKey
#            name: api_key
#            in: header